requirements.txt
redis==3.5.2
cassandra-driver==3.23.0
ujson==2.0.3